Fixes: [link to issue]

Description of the Change
-------------------------

Currently `ConfirmOrphanedSystemResourceJobHandler` deletes all resources it finds to be orphaned which makes it hard to verify correctness without risking resources. This adds a feature flag to the handler to enable deletion on a per-resource-type basis so that we can enable the task for some resources while validating it for others.

Impact
------

Hotfix not needed

Rollback Plan
-------------

If this commit is reverted then we need to validate that the flag in the BE DB to enable `WatchOrphanedSystemResourceJobHandler` is *disabled* as after this change rolls out it will probably be enabled. If this is rolled back when that flag is enabled we would begin deleting all resource types.

+++ src/Codespaces/EnvironmentManager/Src/EnvironmentManager/Handlers/ConfirmOrphanedSystemResourceJobHandler.cs
@@ -3,11 +3,13 @@
 // </copyright>
 
 using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.VsSaaS.Common;
 using Microsoft.VsSaaS.Diagnostics;
 using Microsoft.VsSaaS.Diagnostics.Extensions;
+using Microsoft.VsSaaS.Services.CloudEnvironments.Common.Configuration.KeyGenerator;
 using Microsoft.VsSaaS.Services.CloudEnvironments.Common.Contracts;
 using Microsoft.VsSaaS.Services.CloudEnvironments.Common.Contracts.Constants;
 using Microsoft.VsSaaS.Services.CloudEnvironments.Common.Contracts.Payloads;
@@ -19,16 +21,27 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.EnvironmentManager.Handler
 {
     public class ConfirmOrphanedSystemResourceJobHandler : JobHandlerPayloadBase<ConfirmOrphanedSystemResourcePayload>, IJobHandlerTarget
     {
+        private static readonly ResourceType[] DefaultEnabledResourceTypesToDelete = new[]
+        {
+             ResourceType.ComputeVM,
+             ResourceType.KeyVault,
+        };
+
+        private static readonly string DefaultEnabledResourceTypesToDeleteConfigValue =
+            string.Join(",", DefaultEnabledResourceTypesToDelete.Select((t) => t.ToString()));
+
         public ConfirmOrphanedSystemResourceJobHandler(
             ICloudEnvironmentRepository environmentRepository,
             ISecretStoreRepository secretStoreRepository,
             IResourceBrokerResourcesExtendedHttpContract resourceBroker,
-            IControlPlaneInfo controlPlane)
+            IControlPlaneInfo controlPlane,
+            IConfigurationReader configurationReader)
         {
             EnvironmentRepository = environmentRepository;
             SecretStoreRepository = secretStoreRepository;
             ResourceBroker = resourceBroker;
             ControlPlane = controlPlane;
+            ConfigurationReader = configurationReader;
         }
 
         /// <inheritdoc/>
@@ -48,6 +61,10 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.EnvironmentManager.Handler
 
         private IControlPlaneInfo ControlPlane { get; }
 
+        private IConfigurationReader ConfigurationReader { get; }
+
+        private string ConfigurationBaseName => "ConfirmOrphanedSystemResourceJobHandler";
+
         protected override Task HandleJobAsync(ConfirmOrphanedSystemResourcePayload payload, IDiagnosticsLogger logger, CancellationToken cancellationToken)
         {
             return logger.OperationScopeAsync(
@@ -61,13 +78,21 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.EnvironmentManager.Handler
                         .FluentAddBaseValue("ResourceId", resourceId)
                         .FluentAddBaseValue("ResourceType", resourceType);
 
+                    // Don't pass NewChildLogger here as it will be adding properties to the logger
                     var resourceIsReferenced = await CheckResourceReferencesByTypeAsync(resourceId, resourceType, childLogger);
 
                     childLogger.FluentAddBaseValue("ResourceIsReferenced", resourceIsReferenced);
 
                     if (!resourceIsReferenced)
                     {
-                        await ResourceBroker.DeleteAsync(Guid.Empty, Guid.Parse(resourceId), childLogger.NewChildLogger());                        
+                        var isDeleteEnabled = await CanDeleteResourceType(resourceType, childLogger.NewChildLogger());
+
+                        childLogger.FluentAddBaseValue("ResourceDeleteAttempted", isDeleteEnabled);
+
+                        if (isDeleteEnabled)
+                        {
+                            await ResourceBroker.DeleteAsync(Guid.Empty, Guid.Parse(resourceId), childLogger.NewChildLogger());
+                        }
                     }
                 });
         }
@@ -117,6 +142,23 @@ namespace Microsoft.VsSaaS.Services.CloudEnvironments.EnvironmentManager.Handler
             }
 
             return resourceIsReferenced;
+        }
+
+        private async Task<bool> CanDeleteResourceType(ResourceType resourceType, IDiagnosticsLogger logger)
+        {
+            var enabledResourceTypes = await this.ConfigurationReader.ReadSettingAsync(ConfigurationBaseName, "enabled-resource-types", logger.NewChildLogger(), DefaultEnabledResourceTypesToDeleteConfigValue);
+            if (enabledResourceTypes == default)
+            {
+                // No resource types are enabled
+                return false;
+            }
+
+            var enabledResourceTypeList = enabledResourceTypes.Split(',').Select(type => type.Trim());
+
+            var resourceTypeString = resourceType.ToString();
+            var isEnabled = enabledResourceTypeList.Any((enabledType) => string.Equals(enabledType, resourceTypeString, StringComparison.OrdinalIgnoreCase));
+
+            return isEnabled;
         }
     }
 }
